# openapi.json
{"openapi":"3.1.0","info":{"title":"PII Shield","description":"Intelligent De-identification Service for detecting and anonymizing PII","version":"0.1.0"},"paths":{"/detect":{"post":{"summary":"Detect Pii","description":"Detect PII in text without de-identification.","operationId":"detect_pii_detect_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetectRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DetectResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/anonymize":{"post":{"summary":"Anonymize Text","description":"Detect and de-identify PII in text.","operationId":"anonymize_text_anonymize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnonymizeRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnonymizeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}}}}},"/":{"get":{"summary":"Root","description":"Root endpoint with API info.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AnonymizeRequest":{"properties":{"text":{"type":"string","minLength":1,"title":"Text","description":"Text to anonymize"},"strategy":{"type":"string","title":"Strategy","description":"De-identification strategy: redaction, masking, or hashing","default":"redaction"}},"type":"object","required":["text"],"title":"AnonymizeRequest","description":"Request body for /anonymize endpoint."},"AnonymizeResponse":{"properties":{"original_text":{"type":"string","title":"Original Text"},"processed_text":{"type":"string","title":"Processed Text"},"matches":{"items":{"$ref":"#/components/schemas/PIIMatchSchema"},"type":"array","title":"Matches"},"summary":{"$ref":"#/components/schemas/SummarySchema"},"processing_time_ms":{"type":"number","title":"Processing Time Ms"}},"type":"object","required":["original_text","processed_text","matches","summary","processing_time_ms"],"title":"AnonymizeResponse","description":"Response body for /anonymize endpoint."},"DetectRequest":{"properties":{"text":{"type":"string","minLength":1,"title":"Text","description":"Text to analyze for PII"}},"type":"object","required":["text"],"title":"DetectRequest","description":"Request body for /detect endpoint."},"DetectResponse":{"properties":{"matches":{"items":{"$ref":"#/components/schemas/PIIMatchSchema"},"type":"array","title":"Matches"},"summary":{"$ref":"#/components/schemas/SummarySchema"},"processing_time_ms":{"type":"number","title":"Processing Time Ms"}},"type":"object","required":["matches","summary","processing_time_ms"],"title":"DetectResponse","description":"Response body for /detect endpoint."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"version":{"type":"string","title":"Version"}},"type":"object","required":["status","version"],"title":"HealthResponse","description":"Response body for /health endpoint."},"PIIMatchSchema":{"properties":{"type":{"type":"string","title":"Type"},"text":{"type":"string","title":"Text"},"start":{"type":"integer","title":"Start"},"end":{"type":"integer","title":"End"},"confidence":{"type":"number","title":"Confidence"},"detector":{"type":"string","title":"Detector"}},"type":"object","required":["type","text","start","end","confidence","detector"],"title":"PIIMatchSchema","description":"Schema for a single PII match."},"SummarySchema":{"properties":{"pii_found":{"type":"boolean","title":"Pii Found"},"total_count":{"type":"integer","title":"Total Count"},"by_type":{"additionalProperties":{"type":"integer"},"type":"object","title":"By Type"}},"type":"object","required":["pii_found","total_count","by_type"],"title":"SummarySchema","description":"Summary statistics for detection results."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}